Fix security flaw with delete route

+++ backend/controllers/stuff.js
@@ -59,17 +59,30 @@ exports.modifyThing = (req, res, next) => {
 };
 
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({ _id: req.params.id })
-    .then(() => {
-      res.status(200).json({
-        message: 'Deleted!',
+  Thing.findOne({ _id: req.params.id }).then((thing) => {
+    if (!thing) {
+      return res.status(404).json({
+        error: new Error('Thing not found!'),
       });
-    })
-    .catch((error) => {
-      res.status(400).json({
-        error: error,
+    }
+    if (thing.userId !== req.params.userId) {
+      return res.status(400).json({
+        error: new Error('Unauthorized request!'),
       });
-    });
+    }
+
+    Thing.deleteOne({ _id: req.params.id })
+      .then(() => {
+        res.status(200).json({
+          message: 'Deleted!',
+        });
+      })
+      .catch((error) => {
+        res.status(400).json({
+          error: error,
+        });
+      });
+  });
 };
 
 exports.getAllStuff = (req, res, next) => {
@@ -82,4 +95,4 @@ exports.getAllStuff = (req, res, next) => {
         error: error,
       });
     });
-}
+};
